refactor: nest GraphQL operations under a bulkCreateUsers namespace container
Move the root Query (bulkCreateUsersMaxUploadSize) + root Mutation (bulkCreateUsersImport)
flat fields under one hierarchical namespace container, per the Jahia GraphQL convention.

- Query:    bulkCreateUsersMaxUploadSize -> bulkCreateUsers { maxUploadSize }
- Mutation: bulkCreateUsersImport        -> bulkCreateUsers { importUsers }

New holder types BulkCreateUsersQuery/Mutation carry the operations as instance
@GraphQLField methods (BulkCreateUsersResult return type stays a separate file); the
*Extension classes keep a single bulkCreateUsers accessor field. @GraphQLRequiresPermission
stays on each leaf. The unit test was repointed to the holder (isValidSiteKey moved with the
mutation). Frontend gql.js + jsx and Cypress fixtures/specs updated in lockstep.
