+++ tests/cypress/fixtures/graphql/mutation/importUsers.graphql
@@ -1,10 +1,12 @@
 mutation BulkCreateUsersImport($csvContent: String!, $separator: String, $siteKey: String, $selectedColumns: [String], $overwrite: Boolean) {
-    bulkCreateUsersImport(csvContent: $csvContent, separator: $separator, siteKey: $siteKey, selectedColumns: $selectedColumns, overwrite: $overwrite) {
-        success
-        createdCount
-        updatedCount
-        skippedCount
-        errorCount
-        errors
+    bulkCreateUsers {
+        importUsers(csvContent: $csvContent, separator: $separator, siteKey: $siteKey, selectedColumns: $selectedColumns, overwrite: $overwrite) {
+            success
+            createdCount
+            updatedCount
+            skippedCount
+            errorCount
+            errors
+        }
     }
 }
