+++ tests/cypress/fixtures/graphql/query/maxUploadSize.graphql
@@ -1,3 +1,5 @@
 query BulkCreateUsersMaxUploadSize {
-    bulkCreateUsersMaxUploadSize
+    bulkCreateUsers {
+        maxUploadSize
+    }
 }
